feat: add mac commands

+++ uninstall.sh
@@ -1,4 +1,8 @@
 #!/usr/bin/env bash
 
-# Uninstall htpasswd utility
-sudo apt-get remove -y apache2-utils
+SYSTEM=$3
+
+if [ "$SYSTEM" != "mac" ]; then
+    # Uninstall htpasswd utility
+    sudo apt-get remove -y apache2-utils
+fi
